refactor: split apply-dark-mode hook into multiple hooks
