+++ frontend/src/pages/cheatsheet.tsx
@@ -4,12 +4,12 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { CheatsheetContent } from '../components/editor-page/app-bar/cheatsheet/cheatsheet-content'
-import { useApplyDarkMode } from '../hooks/dark-mode/use-apply-dark-mode'
+import { useApplyDarkModeStyle } from '../hooks/dark-mode/use-apply-dark-mode-style'
 import type { NextPage } from 'next'
 import { Container } from 'react-bootstrap'
 
 const CheatsheetPage: NextPage = () => {
-  useApplyDarkMode()
+  useApplyDarkModeStyle()
 
   return (
     <Container>
